search-embeddings-A/pyproject.toml
[project]
name = "contentful-pinecone-embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "contentful>=2.5.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "numpy>=2.4.2",
    "pinecone-client>=6.0.0",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
]
